+++ lib/usd/hdMaya/adapters/aiSkydomeLightAdapter.cpp
@@ -26,6 +26,7 @@
 #include "lightAdapter.h"
 #include "mayaAttrs.h"
 #include "tokens.h"
+#include "../utils.h"
 
 #include <maya/MPlugArray.h>
 
@@ -94,7 +95,15 @@ public:
         } else if (paramName == HdLightTokens->enableColorTemperature) {
             return VtValue(false);
         } else if (paramName == HdLightTokens->textureResource) {
-            return {};
+            auto fileObj =
+                GetConnectedFileNode(GetNode(), HdMayaAdapterTokens->color);
+            // TODO: Return a default, white texture?
+            // Ideally we would want to return a custom texture resource based
+            // on the color, but not sure how easy that would be.
+            if (fileObj == MObject::kNullObj) { return {}; }
+            return VtValue{GetFileTextureResource(
+                fileObj, GetFileTexturePath(MFnDependencyNode(fileObj)),
+                GetDelegate()->GetParams().textureMemoryPerTexture)};
         }
         return {};
     }
